Fonctionnalité #5307 ajout du champ SISERV à la consultation et modification modification du reporting maven pour le site

+++ src/main/java/nc/noumea/mairie/ads/dto/ReferenceDto.java
@@ -2,14 +2,14 @@ package nc.noumea.mairie.ads.dto;
 
 public class ReferenceDto {
 
-	public long id;
+	public Integer id;
 	public String label;
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
